pyproject.toml
[project]
name = "email-schedule-send-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.10.5",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
]
